Update all dependencies
Use wildcard dependencies instead of strict ones. It provides more
compatibility with other addons and apps. Especially it is important
for "dependencies" section.

Update all the dependencies to latest. remove ember-cli-update because
it is now part of ember-cli itself.

Update test helper 'upload' to be coble with the newest ember
test-helpers.

Add exceptions for ember/no-observers rule, because it is in recommended
set of rules now.